Distribute code into layer and neuron packages
