+++ captum/attr/__init__.py
@@ -1,18 +1,26 @@
 #!/usr/bin/env python3
 
 from ._core.integrated_gradients import IntegratedGradients  # noqa
-from ._core.internal_influence import InternalInfluence  # noqa
 from ._core.deep_lift import DeepLift, DeepLiftShap  # noqa
 from ._core.input_x_gradient import InputXGradient  # noqa
 from ._core.saliency import Saliency  # noqa
 from ._core.noise_tunnel import NoiseTunnel  # noqa
 from ._core.gradient_shap import GradientShap  # noqa
-from ._core.layer_conductance import LayerConductance  # noqa
-from ._core.layer_gradient_x_activation import LayerGradientXActivation  # noqa
-from ._core.layer_activation import LayerActivation  # noqa
-from ._core.neuron_conductance import NeuronConductance  # noqa
-from ._core.neuron_gradient import NeuronGradient  # noqa
-from ._core.neuron_integrated_gradients import NeuronIntegratedGradients  # noqa
+from ._core.guided_backprop_deconvnet import GuidedBackprop, Deconvolution # noqa
+from ._core.guided_grad_cam import GuidedGradCam # noqa
+from ._core.feature_ablation import FeatureAblation # noqa
+from ._core.layer.layer_conductance import LayerConductance  # noqa
+from ._core.layer.layer_gradient_x_activation import LayerGradientXActivation  # noqa
+from ._core.layer.layer_activation import LayerActivation  # noqa
+from ._core.layer.internal_influence import InternalInfluence  # noqa
+from ._core.layer.grad_cam import LayerGradCam # noqa
+from ._core.layer.layer_deep_lift import LayerDeepLift, LayerDeepLiftShap # noqa
+from ._core.neuron.neuron_conductance import NeuronConductance  # noqa
+from ._core.neuron.neuron_gradient import NeuronGradient  # noqa
+from ._core.neuron.neuron_integrated_gradients import NeuronIntegratedGradients  # noqa
+from ._core.neuron.neuron_deep_lift import NeuronDeepLift, NeuronDeepLiftShap # noqa
+from ._core.neuron.neuron_guided_backprop_deconvnet import NeuronDeconvolution, NeuronGuidedBackprop # noqa
+
 from ._models.base import (
     InterpretableEmbeddingBase,
     TokenReferenceBase,
@@ -35,12 +43,23 @@ __all__ = [
     "DeepLift",
     "InputXGradient",
     "Saliency",
+    "GuidedBackprop",
+    "Deconvolution",
+    "GuidedGradCam",
+    "FeatureAblation",
     "LayerConductance",
     "LayerGradientXActivation",
     "LayerActivation",
+    "LayerGradCam",
+    "LayerDeepLift",
+    "LayerDeepLiftShap",
     "NeuronConductance",
     "NeuronGradient",
     "NeuronIntegratedGradients",
+    "NeuronDeepLift",
+    "NeuronDeepLiftShap",
+    "NeuronDeconvolution",
+    "NeuronGuidedBackprop",
     "NoiseTunnel",
     "GradientShap",
     "InterpretableEmbeddingBase",
