+++ captum/attr/_core/layer/grad_cam.py
@@ -1,25 +1,24 @@
 #!/usr/bin/env python3
-from .._utils.attribution import LayerAttribution
-from .._utils.common import _format_input, _format_additional_forward_args
-from .._utils.gradient import compute_layer_gradients_and_eval
+import torch
+
+from ..._utils.attribution import LayerAttribution
+from ..._utils.common import _format_input, _format_additional_forward_args
+from ..._utils.gradient import compute_layer_gradients_and_eval
 
 
-class LayerGradientXActivation(LayerAttribution):
+class LayerGradCam(LayerAttribution):
     def __init__(self, forward_func, layer, device_ids=None):
         r"""
-        Args:
+        Args
 
             forward_func (callable):  The forward function of the model or any
                           modification of it
             layer (torch.nn.Module): Layer for which attributions are computed.
-                          Output size of attribute matches this layer's input or
-                          output dimensions, depending on whether we attribute to
-                          the inputs or outputs of the layer, corresponding to
-                          attribution of each neuron in the input or output of
-                          this layer.
-                          Currently, it is assumed that the inputs or the outputs
-                          of the layer, depending on which one is used for
-                          attribution, can only be a single tensor.
+                          Output size of attribute matches this layer's output
+                          dimensions, except for dimension 2, which will be 1,
+                          since GradCAM sums over channels.
+                          Currently, only layers with a single tensor output are
+                          supported.
             device_ids (list(int)): Device ID list, necessary only if forward_func
                           applies a DataParallel model. This allows reconstruction of
                           intermediate outputs from batched results across devices.
@@ -36,8 +35,35 @@ class LayerGradientXActivation(LayerAttribution):
         attribute_to_layer_input=False,
     ):
         r"""
-            Computes element-wise product of gradient and activation for selected
-            layer on given inputs.
+            Computes GradCAM attribution for chosen layer. GradCAM is designed for
+            convolutional neural networks, and is usually applied to the last
+            convolutional layer.
+
+            GradCAM computes the gradients of the target output with respect to
+            the given layer, averages for each output channel (dimension 2 of
+            output), and multiplies the average gradient for each channel by the
+            layer activations. The results are summed over all channels.
+
+            Note that in the original GradCAM algorithm described in the paper,
+            ReLU is applied to the output, returning only non-negative attributions.
+            For providing more flexibility to the user, we choose to not perform the
+            ReLU internally and return the sign information. To match the original
+            GradCAM algorithm, it is necessary to perform ReLU on the returned
+            attributions or alternatively only visualize the positive attributions.
+
+            Note: this procedure sums over the second dimension (# of channels),
+            so the output of GradCAM attributions will have a second
+            dimension of 1, but all other dimensions will match that of the layer
+            output.
+
+            GradCAM attributions are generally upsampled and can be viewed as a
+            mask to the input, since a convolutional layer output generally
+            matches the input image spatially. This upsampling can be performed
+            using LayerAttribution.interpolate, as shown in the example below.
+
+            More details regarding the GradCAM method can be found in the
+            original paper here:
+            https://arxiv.org/pdf/1610.02391.pdf
 
             Args:
 
@@ -87,37 +113,45 @@ class LayerGradientXActivation(LayerAttribution):
                             to these arguments.
                             Default: None
                 attribute_to_layer_input (bool, optional): Indicates whether to
-                            compute the attribution with respect to the layer input
+                            compute the attributions with respect to the layer input
                             or output. If `attribute_to_layer_input` is set to True
-                            then the attributions will be computed with respect to
+                            then the attributions will be computed with respect to the
                             layer input, otherwise it will be computed with respect
                             to layer output.
                             Note that currently it is assumed that either the input
-                            or the output of internal layer, depending on whether we
-                            attribute to the input or output, is a single tensor.
+                            or the outputs of internal layers, depending on whether we
+                            attribute to the input or output, are single tensors.
                             Support for multiple tensors will be added later.
                             Default: False
 
             Returns:
                 *tensor* of **attributions**:
                 - **attributions** (*tensor*):
-                            Product of gradient and activation for each
-                            neuron in given layer output.
-                            Attributions will always be the same size as the
-                            output of the given layer.
+                            Attributions based on GradCAM method.
+                            Attributions will be the same size as the
+                            output of the given layer, except for dimension 2,
+                            which will be 1 due to summing over channels.
 
             Examples::
 
                 >>> # ImageClassifier takes a single input tensor of images Nx3x32x32,
                 >>> # and returns an Nx10 tensor of class probabilities.
-                >>> # It contains an attribute conv1, which is an instance of nn.conv2d,
-                >>> # and the output of this layer has dimensions Nx12x32x32.
+                >>> # It contains a layer conv4, which is an instance of nn.conv2d,
+                >>> # and the output of this layer has dimensions Nx50x8x8.
+                >>> # It is the last convolution layer, which is the recommended
+                >>> # use case for GradCAM.
                 >>> net = ImageClassifier()
-                >>> layer_ga = LayerGradientXActivation(net, net.conv1)
+                >>> layer_gc = LayerGradCam(net, net.conv4)
                 >>> input = torch.randn(2, 3, 32, 32, requires_grad=True)
-                >>> # Computes layer activation x gradient for class 3.
-                >>> # attribution size matches layer output, Nx12x32x32
-                >>> attribution = layer_ga.attribute(input, 3)
+                >>> # Computes layer GradCAM for class 3.
+                >>> # attribution size matches layer output except for dimension
+                >>> # 1, so dimensions of attr would be Nx1x8x8.
+                >>> attr = layer_gc.attribute(input, 3)
+                >>> # GradCAM attributions are often upsampled and viewed as a
+                >>> # mask to the input, since the convolutional layer output
+                >>> # spatially matches the original input image.
+                >>> # This can be done with LayerAttribution's interpolate method.
+                >>> upsampled_attr = LayerAttribution.interpolate(attr, (32, 32))
         """
         inputs = _format_input(inputs)
         additional_forward_args = _format_additional_forward_args(
@@ -134,4 +168,12 @@ class LayerGradientXActivation(LayerAttribution):
             device_ids=self.device_ids,
             attribute_to_layer_input=attribute_to_layer_input,
         )
-        return layer_gradients * layer_eval
+        summed_grads = torch.mean(
+            layer_gradients,
+            dim=tuple(x for x in range(2, len(layer_gradients.shape))),
+            keepdim=True,
+        )
+
+        non_neg_scaled_act = torch.sum(summed_grads * layer_eval, dim=1, keepdim=True)
+
+        return non_neg_scaled_act
