+++ captum/attr/_core/layer/layer_activation.py
@@ -1,8 +1,8 @@
 #!/usr/bin/env python3
 import torch
 
-from .._utils.attribution import LayerAttribution
-from .._utils.gradient import _forward_layer_eval
+from ..._utils.attribution import LayerAttribution
+from ..._utils.gradient import _forward_layer_eval
 
 
 class LayerActivation(LayerAttribution):
