+++ tests/attr/neuron/test_neuron_integrated_gradients.py
@@ -4,14 +4,14 @@ import unittest
 
 import torch
 from captum.attr._core.integrated_gradients import IntegratedGradients
-from captum.attr._core.neuron_integrated_gradients import NeuronIntegratedGradients
+from captum.attr._core.neuron.neuron_integrated_gradients import NeuronIntegratedGradients
 
-from .helpers.basic_models import (
+from ..helpers.basic_models import (
     BasicModel_ConvNet,
     BasicModel_MultiLayer,
     BasicModel_MultiLayer_MultiInput,
 )
-from .helpers.utils import assertArraysAlmostEqual, BaseTest
+from ..helpers.utils import assertArraysAlmostEqual, BaseTest
 
 
 class Test(BaseTest):
