+++ tests/attr/test_deconvolution.py
@@ -4,7 +4,7 @@ import unittest
 
 import torch
 from captum.attr._core.guided_backprop_deconvnet import Deconvolution
-from captum.attr._core.neuron_guided_backprop_deconvnet import NeuronDeconvolution
+from captum.attr._core.neuron.neuron_guided_backprop_deconvnet import NeuronDeconvolution
 
 from .helpers.basic_models import BasicModel_ConvNet_One_Conv
 from .helpers.utils import assertTensorAlmostEqual, BaseTest
